pyproject.toml
[tool.poetry]
name = "amazon-price-tracker"
version = "0.1.0"
description = "Script to track price of an amazon product and notify via email"
authors = ["Fausto Ramos <fausto.ramosr92@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.17.1"
bs4 = "^0.0.1"
requests = "^2.25.1"
lxml = "^4.6.3"
selenium = "^3.141.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.5b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
